change a lot
Change load state to message/info
Add speed control on game pause menu
Change msg function to a class
the image was a bad idea, will probably remove it later (was for buff, when ball move over it. 3 buff (e.g. star) permit destroy a lightning which suppose to partially destroy the shield arrow the heart. Taking the heart end the stage. <= NOT DONE.